// praxis-web/src/app/schemas/student.ts
export class Student {
    name: string;
    birthdate: string;
    email: string;
    password: string;
    typeOfPraxis: string;
    biography: string;
    urlvideo: string;
    documentType: string; // Español o ingles ?
    documentNumber: number;
    university: string; //Solo mostrar u nacional en la lista.
    semester: number;
    
}